Reorganizing the file structure a little
