+++ src/unity/toolkits/object_detection/one_shot_object_detection/util/mapping_function.hpp
